+++ changelog.md
@@ -1,3 +1,9 @@
+# 1.4.3
+
+- Fixed Bass being boosted on some Samsung devices
+- Added Trail and Wave Trail options to icon effects
+- Added option to customize the colours used in the fade icon effect mode
+
 # 1.4.2
 
 - Added Individual X/Y scale mode in Creator Set Scale